requirements.txt
asgiref==3.11.1
click==8.3.1
colorama==0.4.6
croniter==6.0.0
Django==6.0.2
django-cors-headers==4.9.0
django-redis==6.0.0
django-rq==3.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.2.0
rq==2.6.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
